Updated the engine and gameManager. Engine and GameManager now have a start game function that loads the default room. Rooms now have a musicName property, which plays music once the room is loaded. Added draw and drawGui methods to GameManager. Engine also now stores event listener IDs and cleans them up when the engine is closed. Added a closeEngine method to the Engine which properly does clean up.

+++ include/engine/main_stuff/audioController.hpp
@@ -50,10 +50,10 @@ class AudioController
         bool playingMusic();
 
         /// @return The number of loaded music assets.
-        int getMusicCount();
+        size_t getMusicCount();
 
         /// @return The number of loaded sound assets.
-        int getSoundCount();
+        size_t getSoundCount();
 
         #pragma endregion Getters
